fix: mudança do tipo de pk

+++ src/main/java/com/futureh/dronefeeder/infrastructure/persistence/hibernate/repository/entrega/EntregaRepository.java
@@ -1,9 +1,9 @@
 package com.futureh.dronefeeder.infrastructure.persistence.hibernate.repository.entrega;
 
-import org.springframework.data.repository.CrudRepository;
-
 import com.futureh.dronefeeder.domain.entrega.model.Entrega;
 
-public interface EntregaRepository extends CrudRepository<Entrega, Long> {
+import org.springframework.data.repository.CrudRepository;
+
+public interface EntregaRepository extends CrudRepository<Entrega, Integer> {
 
 }
